# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 62782285c97c70218d654808
  title: j2ee
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /livre:
    post:
      x-skaffolder-id: 627827a1c97c70218d654842
      x-skaffolder-name: create
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - livre
      parameters:
        - name: obj
          x-skaffolder-type: livre
          in: path
          description: Object to insert
          required: true
          schema:
            type: livre
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/livre"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                auteur:
                  type: string
                titre:
                  type: string
    get:
      x-skaffolder-id: 627827b6c97c70218d654848
      x-skaffolder-name: list
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF livre
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - livre
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/livre"
  "/livre/{id}":
    delete:
      x-skaffolder-id: 627827acc97c70218d654844
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - livre
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id livre
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      x-skaffolder-id: 627827b2c97c70218d654846
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - livre
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id livre
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 627827bbc97c70218d654849
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: livre
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - livre
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id livre
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/livre"
    post:
      x-skaffolder-id: 627829c0c97c70218d654867
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6278278cc97c70218d654840
      x-skaffolder-resource: livre
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - livre
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/livre"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/livre"
  /pret:
    post:
      x-skaffolder-id: 6278283cc97c70218d654852
      x-skaffolder-name: create
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - pret
      parameters:
        - name: obj
          x-skaffolder-type: pret
          in: path
          description: Object to insert
          required: true
          schema:
            type: pret
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pret"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: integer
                date_retour:
                  type: integer
                id_adherent:
                  type: number
                id_livre:
                  type: number
    get:
      x-skaffolder-id: 62782849c97c70218d654858
      x-skaffolder-name: list
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF pret
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - pret
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/pret"
  "/pret/{id}":
    delete:
      x-skaffolder-id: 62782840c97c70218d654854
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - pret
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id pret
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6278284cc97c70218d654859
      x-skaffolder-name: get
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: pret
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - pret
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id pret
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pret"
    post:
      x-skaffolder-id: 62782844c97c70218d654856
      x-skaffolder-name: update
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: pret
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - pret
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id pret
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/pret"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/pret"
  "/pret/findByadherent_pret/{key}":
    get:
      x-skaffolder-id: 62782990c97c70218d654863
      x-skaffolder-name: findByadherent_pret
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: findByadherent_pret
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByadherent_pret
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByadherent_pret/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByadherent_pret
      tags:
        - pret
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/pret"
  "/pret/findBydate/{key}":
    get:
      x-skaffolder-id: 6278285bc97c70218d65485b
      x-skaffolder-name: findBydate
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: findBydate
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBydate
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBydate/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBydate
      tags:
        - pret
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource date to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/pret"
  "/pret/findBylivre_pret/{key}":
    get:
      x-skaffolder-id: 627829c0c97c70218d654869
      x-skaffolder-name: findBylivre_pret
      x-skaffolder-id-resource: 62782805c97c70218d654850
      x-skaffolder-resource: pret
      x-skaffolder-crudAction: findBylivre_pret
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBylivre_pret
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBylivre_pret/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBylivre_pret
      tags:
        - pret
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/pret"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    livre:
      x-skaffolder-id: 6278278cc97c70218d654840
      x-skaffolder-id-db: 62782285c97c70218d654809
      x-skaffolder-id-entity: 6278278cc97c70218d65483d
      x-skaffolder-url: /livre
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        auteur:
          type: string
          x-skaffolder-id-attr: 6278278cc97c70218d65483f
          x-skaffolder-type: String
        titre:
          type: string
          x-skaffolder-id-attr: 6278278cc97c70218d65483e
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    pret:
      x-skaffolder-id: 62782805c97c70218d654850
      x-skaffolder-id-db: 62782285c97c70218d654809
      x-skaffolder-id-entity: 62782805c97c70218d65484d
      x-skaffolder-url: /pret
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        date:
          type: integer
          x-skaffolder-id-attr: 62782805c97c70218d65484e
          x-skaffolder-type: Date
        date_retour:
          type: integer
          x-skaffolder-id-attr: 62782805c97c70218d65484f
          x-skaffolder-type: Date
        id_adherent:
          type: number
          x-skaffolder-id-attr: 62782895c97c70218d65485e
          x-skaffolder-type: Number
        id_livre:
          type: number
          x-skaffolder-id-attr: 62782895c97c70218d65485d
          x-skaffolder-type: Number
      x-skaffolder-relations:
        adherent_pret:
          x-skaffolder-id: 62782973c97c70218d654860
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 62782805c97c70218d65484d
          x-skaffolder-ent2: 627826cac97c70218d654828
        livre_pret:
          x-skaffolder-id: 62782973c97c70218d65485f
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 62782805c97c70218d65484d
          x-skaffolder-ent2: 6278278cc97c70218d65483d
      required:
        - _id
...